[REF] use monkey patching for best support of inheriting
